Update to v2.7.1.0
* Change 'gallery type' to 'gallery focus'. Rather than forcing a
* gallery to only display one content type in the main and lightbox
* areas with all other content as supporting files, gallery focus lets
* you set a default focus for the gallery, and then the user viewing is
* able to toggle between audio/image/video being the focus.

* Tags: You'll be able to add tags to collections, galleries and items.

* Collection types: a collection activity will be able to be set to 1 of
* 4 types.
** Instructor collection - only instructors can create galleries/items.
** Contributed collection - all users can create galleries/items - but
this collection can't be used with the assign submission plugin.
** Assignment collection - users can only see their own galleries - and
can be used for an assign submission.
** Peer reviewed collection - users can see each others galleries - and
can be used for an assign submission.

* Gallery modes. Two modes you can pick from when making a gallery.
** Standard - the same as now.
** YouTube - only accepts youtube videos as items.

+++ rest.php
@@ -59,7 +59,7 @@ switch($requestmethod) {
             $info = $object->get_socialinfo();
             echo json_encode($info);
         } else if ($action == 'get_sample_targets') {
-            $info = mediagallery_get_sample_targets($course);
+            $info = mediagallery_get_sample_targets($course, $object);
             echo json_encode($info);
         } else if ($action == 'embed') {
             $output = $PAGE->get_renderer('mod_mediagallery')->embed_html($object);
@@ -71,6 +71,7 @@ switch($requestmethod) {
             $data->id = isset($matches[2]) ? $matches[2] : null;
             $data->url = isset($matches[3]) ? str_replace('\/', '/', $matches[3]) : null;
             $data->type = isset($matches[1]) ? $matches[1] : $object->type(true);
+            $data->objectid = $object->objectid;
             $data->flow = !empty($matches[1]);
             echo json_encode($data);
         } else if ($action == 'metainfo') {
@@ -94,7 +95,26 @@ switch($requestmethod) {
     break;
 
     case 'DELETE':
-        if ($object->delete()) {
+        $success = false;
+        $isowner = $object->is_thebox_creator_or_agent();
+
+        if ($class == 'collection') {
+            if (!has_capability('mod/mediagallery:manage', $object->get_context())) {
+                throw new moodle_exception("no permission");
+            }
+            $success = true;
+        } else {
+            if (empty($object->objectid)) {
+                // Non-box.
+                if ($object->delete()) {
+                    $success = true;
+                }
+            } else {
+                $success = true;
+            }
+
+        }
+        if ($success) {
             echo json_encode('success');
         } else {
             throw new moodle_exception("failed to delete $class $id");
